[proliferate] split all changes frame
